// src/api/generated/models/ListComputeSettingsInputDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ListComputeSettingsInputValueDto } from './ListComputeSettingsInputValueDto';

export type ListComputeSettingsInputDto = {
    inputSettingTypeId: string;
    name: string;
    values: Array<ListComputeSettingsInputValueDto>;
};